# cgen: cleanup cgen expr writing