@spytheman 2024-12-11
This has non trivial implications, because g.write and g.writeln are optimized, when their argument is a string interpolation, into several g.write() calls, that do not use string interpolation at all.

With the code of the PR, calls to `g.write_expr()` will not be, and when called with interpolated strings, the interpolation will remain, and will be evaluated, before passing the result to `g.write_expr()`.

For places, that are not called frequently, that does not matter, and can be beneficial, since there will be less work for the transformer, and the produced C code will be a bit smaller, but for hot spots, it can increase the RAM and CPU usage of cgen a lot.

My concern is mostly, that since the PR introduces those additional write methods, that are not optimized by the transform stage (and I do not see an easy way for it to do so, without special casing them), it creates the risk of migrating more places in cgen to them, little by little losing the advantage that 3e6fc36 from 2024-10-09 gave.

Here is a comparison between the output of cgen on master and here:
![image](https://github.com/user-attachments/assets/a54c0cd3-7c9d-4db7-bf84-71e6589dabab)

As you can see, the str_intp() call is avoided on master, but is present here, which means an overhead of 2 allocations + 2 copies, vs just calling .write() more times, with simpler arguments static string literals arguments, with no additional allocations.

That said, currently the balance is slightly in favor of this PR, judged by the output of `v run .github/workflows/compare_pr_to_master.v`:
```
#0 06:36:38 ^ cleanup_cgen_expr_write /space/v/oo>./v run .github/workflows/compare_pr_to_master.v 
==================================================================================================
Current git branch: cleanup_cgen_expr_write, commit: 6d69e1f
    Compiling new V executables from PR branch: cleanup_cgen_expr_write, commit: 6d69e1f ...
CPU: 2.72s      Real: 2.50s     Elapsed: 0:02.50        RAM: 427208KB   ./v -o vnew1 cmd/v
CPU: 2.74s      Real: 2.49s     Elapsed: 0:02.49        RAM: 427348KB   ./vnew1 -o vnew2 cmd/v
CPU: 2.55s      Real: 2.77s     Elapsed: 0:02.77        RAM: 410676KB   ./vnew2 -no-parallel -o vnew cmd/v
CPU: 0.16s      Real: 0.18s     Elapsed: 0:00.18        RAM: 44064KB    ./vnew -no-parallel -o nhw_current.c examples/hello_world.v
CPU: 2.13s      Real: 2.33s     Elapsed: 0:02.33        RAM: 410768KB   ./vnew -no-parallel -o nv_current.c cmd/v
>Size of        nhw_current.c:      45394
>Size of         nv_current.c:    6122969
>Size of                 vnew:    9478936
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
==================================================================================================
    Compiling old V executables from branch: master, commit: 23bcd91 ...
CPU: 2.75s      Real: 2.51s     Elapsed: 0:02.51        RAM: 428200KB   ./v -o vold1 cmd/v
CPU: 2.77s      Real: 2.54s     Elapsed: 0:02.54        RAM: 428264KB   ./vold1 -o vold2 cmd/v
CPU: 2.52s      Real: 2.76s     Elapsed: 0:02.76        RAM: 411576KB   ./vold2 -no-parallel -o vold cmd/v
CPU: 0.16s      Real: 0.18s     Elapsed: 0:00.18        RAM: 44120KB    ./vold -no-parallel -o ohw_master.c examples/hello_world.v
CPU: 2.15s      Real: 2.36s     Elapsed: 0:02.36        RAM: 411488KB   ./vold -no-parallel -o ov_master.c cmd/v
>Size of         ohw_master.c:      45398
>Size of          ov_master.c:    6129231
>Size of                 vold:    9471776
==================================================================================================
File sizes so far ...
>>>>>> size("  nhw_current.c") - size("   ohw_master.c") =      45394 -      45398 =         -4
>>>>>> size("   nv_current.c") - size("    ov_master.c") =    6122969 -    6129231 =      -6262
>>>>>> size("           vnew") - size("           vold") =    9478936 -    9471776 =       7160
Switched to branch 'cleanup_cgen_expr_write'
==================================================================================================
    Measuring at PR branch: cleanup_cgen_expr_write, commit: 6d69e1f ...
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -0.7% 1.01x faster 112.7ms ± σ:   0.3ms, 112.4ms…113.1ms `./vnew -check-syntax           examples/hello_world.v`
 >  2         base         113.4ms ± σ:   0.3ms, 113.2ms…113.8ms `./vold -check-syntax           examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -0.6% 1.01x faster 111.9ms ± σ:   0.4ms, 111.4ms…112.4ms `./vnew -check-syntax           examples/hello_world.v`
 >  2         base         112.6ms ± σ:   0.2ms, 112.3ms…112.8ms `./vold -check-syntax           examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -0.2% 1.00x faster 113.0ms ± σ:   0.0ms, 113.0ms…113.0ms `./vnew -check-syntax           examples/hello_world.v`
 >  2         base         113.2ms ± σ:   0.2ms, 113.0ms…113.4ms `./vold -check-syntax           examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -0.0% 1.00x ~same~ 166.7ms ± σ:   0.5ms, 166.0ms…167.1ms `./vnew -check                  examples/hello_world.v`
 >  2         base         166.7ms ± σ:   0.6ms, 166.0ms…167.2ms `./vold -check                  examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -0.1% 1.00x ~same~ 167.2ms ± σ:   0.0ms, 167.2ms…167.3ms `./vnew -check                  examples/hello_world.v`
 >  2         base         167.3ms ± σ:   1.2ms, 165.8ms…168.7ms `./vold -check                  examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -0.2% 1.00x faster 166.6ms ± σ:   0.1ms, 166.5ms…166.8ms `./vnew -check                  examples/hello_world.v`
 >  2         base         167.0ms ± σ:   1.0ms, 165.5ms…167.9ms `./vold -check                  examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
    1   -1.3% 1.01x faster 186.4ms ± σ:   0.3ms, 186.0ms…186.8ms `./vnew -no-parallel -o nhw.c   examples/hello_world.v`
 >  2         base         188.9ms ± σ:   0.5ms, 188.3ms…189.3ms `./vold -no-parallel -o ohw.c   examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -1.1% 1.01x faster 187.0ms ± σ:   0.7ms, 186.3ms…187.9ms `./vnew -no-parallel -o nhw.c   examples/hello_world.v`
 >  2         base         189.1ms ± σ:   1.7ms, 186.9ms…191.1ms `./vold -no-parallel -o ohw.c   examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -1.7% 1.02x faster 185.2ms ± σ:   0.7ms, 184.5ms…186.2ms `./vnew -no-parallel -o nhw.c   examples/hello_world.v`
 >  2         base         188.4ms ± σ:   0.1ms, 188.3ms…188.5ms `./vold -no-parallel -o ohw.c   examples/hello_world.v`
>>>>>> size("          nhw.c") - size("          ohw.c") =      45394 -      45398 =         -4
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
 >  1         base         198.7ms ± σ:   2.2ms, 196.6ms…201.7ms `./vold -no-parallel -o ohw.exe examples/hello_world.v`
    2   +0.5% 1.01x slower 199.7ms ± σ:   2.1ms, 196.8ms…201.4ms `./vnew -no-parallel -o nhw.exe examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
    1   -0.6% 1.01x faster 197.9ms ± σ:   0.8ms, 196.9ms…198.7ms `./vnew -no-parallel -o nhw.exe examples/hello_world.v`
 >  2         base         199.2ms ± σ:   1.2ms, 198.0ms…200.8ms `./vold -no-parallel -o ohw.exe examples/hello_world.v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3                                                                                  
    1   -0.4% 1.00x faster 197.1ms ± σ:   0.4ms, 196.6ms…197.4ms `./vnew -no-parallel -o nhw.exe examples/hello_world.v`
 >  2         base         198.0ms ± σ:   1.2ms, 196.2ms…198.9ms `./vold -no-parallel -o ohw.exe examples/hello_world.v`
>>>>>> size("        nhw.exe") - size("        ohw.exe") =     242344 -     242344 =          0
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1                                                                                  
 >  1         base         920.0ms ± σ:   3.6ms, 915.0ms…923.4ms `./vold -check-syntax           cmd/v`
    2   +0.1% 1.00x ~same~ 920.5ms ± σ:   4.7ms, 914.7ms…926.2ms `./vnew -check-syntax           cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2                                                                                  
 >  1         base         912.2ms ± σ:   0.8ms, 911.1ms…913.2ms `./vold -check-syntax           cmd/v`
    2   +1.7% 1.02x slower 927.9ms ± σ:   2.9ms, 924.0ms…930.9ms `./vnew -check-syntax           cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3
 >  1         base         921.9ms ± σ:   5.2ms, 915.2ms…927.8ms `./vold -check-syntax           cmd/v`
    2   +0.4% 1.00x slower 925.6ms ± σ:   4.0ms, 920.6ms…930.5ms `./vnew -check-syntax           cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1
    1   -6.7% 1.07x faster 1571.3ms ± σ:  12.1ms, 1554.4ms…1582.0ms `./vnew -check                  cmd/v`
 >  2         base         1683.7ms ± σ:  63.4ms, 1619.1ms…1769.9ms `./vold -check                  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2
    1   -1.7% 1.02x faster 1591.8ms ± σ:   1.6ms, 1590.0ms…1593.9ms `./vnew -check                  cmd/v`
 >  2         base         1619.6ms ± σ:  49.7ms, 1561.9ms…1683.2ms `./vold -check                  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3
 >  1         base         1574.1ms ± σ:   5.0ms, 1567.9ms…1580.3ms `./vold -check                  cmd/v`
    2   +0.0% 1.00x ~same~ 1574.5ms ± σ:   2.6ms, 1571.5ms…1577.7ms `./vnew -check                  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1
 >  1         base         2354.3ms ± σ:   4.9ms, 2347.7ms…2359.4ms `./vold -no-parallel -o ov.c    cmd/v`
    2   +0.8% 1.01x slower 2372.1ms ± σ:  16.8ms, 2357.8ms…2395.7ms `./vnew -no-parallel -o nv.c    cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2
    1   -0.8% 1.01x faster 2346.4ms ± σ:  12.8ms, 2328.3ms…2356.0ms `./vnew -no-parallel -o nv.c    cmd/v`
 >  2         base         2365.6ms ± σ:  14.5ms, 2345.6ms…2379.4ms `./vold -no-parallel -o ov.c    cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3
 >  1         base         2335.8ms ± σ:  12.8ms, 2318.1ms…2347.8ms `./vold -no-parallel -o ov.c    cmd/v`
    2   +0.2% 1.00x slower 2340.4ms ± σ:   8.2ms, 2329.2ms…2348.9ms `./vnew -no-parallel -o nv.c    cmd/v`
>>>>>> size("           nv.c") - size("           ov.c") =    6122969 -    6123230 =       -261
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  1
 >  1         base         2754.2ms ± σ:   5.3ms, 2747.1ms…2760.0ms `./vold -no-parallel -o ov.exe  cmd/v`
    2   +0.2% 1.00x slower 2760.2ms ± σ:   2.8ms, 2756.4ms…2763.1ms `./vnew -no-parallel -o nv.exe  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  2
    1   -0.1% 1.00x ~same~ 2758.6ms ± σ:   5.1ms, 2751.4ms…2763.2ms `./vnew -no-parallel -o nv.exe  cmd/v`
 >  2         base         2761.1ms ± σ:  16.7ms, 2739.5ms…2780.1ms `./vold -no-parallel -o ov.exe  cmd/v`
Summary after 1 series x 10 runs (%s are relative to first command, or `base`), discard maxs:  7, repeat:  3
 >  1         base         2757.2ms ± σ:   6.1ms, 2752.8ms…2765.8ms `./vold -no-parallel -o ov.exe  cmd/v`
    2   +0.1% 1.00x ~same~ 2760.3ms ± σ:  11.3ms, 2744.2ms…2768.7ms `./vnew -no-parallel -o nv.exe  cmd/v`
>>>>>> size("         nv.exe") - size("         ov.exe") =    9478952 -    9478952 =          0
Done. Total time: 643.417652425 s.
==================================================================================================
Final summary for file diff sizes on their own branches:
>>>>>> size("  nhw_current.c") - size("   ohw_master.c") =      45394 -      45398 =         -4
>>>>>> size("   nv_current.c") - size("    ov_master.c") =    6122969 -    6129231 =      -6262
>>>>>> size("           vnew") - size("           vold") =    9478936 -    9471776 =       7160
==================================================================================================
Final summary for file diff sizes for generated files on the *current* branch:
>>>>>> size("          nhw.c") - size("          ohw.c") =      45394 -      45398 =         -4
>>>>>> size("           nv.c") - size("           ov.c") =    6122969 -    6123230 =       -261
>>>>>> size("        nhw.exe") - size("        ohw.exe") =     242344 -     242344 =          0
>>>>>> size("         nv.exe") - size("         ov.exe") =    9478952 -    9478952 =          0
#0 06:47:27 ^ cleanup_cgen_expr_write /space/v/oo>
```

I am still waiting for the same comparison, but with `-prod`.